+++ app/Resources/views/Text/edit.html.twig
@@ -74,9 +74,3 @@
         'submit_button_value': 'app.save'|trans,
     } %}
 {% endblock %}
-
-{% block javascripts %}
-    {{ parent() }}
-
-    <script src="{{ asset('build/app_text_edit.js') }}"></script>
-{% endblock %}
